Move helpers out of magpietts_inference.py inference example
