+++ examples/tts/magpietts_inference.py
@@ -57,19 +57,13 @@ from __future__ import annotations
 import argparse
 import json
 import os
-import random
 import shutil
-import time
-from dataclasses import fields
 from pathlib import Path
 from typing import List, Optional, Tuple
 
 import numpy as np
-import torch
 
 from nemo.collections.asr.parts.utils.manifest_utils import read_manifest
-from nemo.collections.tts.models.easy_magpietts_inference import EasyModelInferenceParameters
-from nemo.collections.tts.models.magpietts import ModelInferenceParameters
 from nemo.collections.tts.modules.magpietts_inference.evaluate_generated_audio import load_evalset_config
 from nemo.collections.tts.modules.magpietts_inference.evaluation import (
     DEFAULT_VIOLIN_METRICS,
@@ -77,477 +71,33 @@ from nemo.collections.tts.modules.magpietts_inference.evaluation import (
     compute_mean_with_confidence_interval,
     evaluate_generated_audio_dir,
 )
-from nemo.collections.tts.modules.magpietts_inference.inference import (
-    BaseInferenceConfig,
-    BaseInferenceRunner,
-    EasyMagpieInferenceConfig,
-    EasyMagpieInferenceRunner,
-    EasyMagpieMultiturnUserAudioInferenceConfig,
-    EasyMagpieMultiturnUserAudioInferenceRunner,
-    MagpieInferenceConfig,
-    MagpieInferenceRunner,
-)
+from nemo.collections.tts.modules.magpietts_inference.inference import BaseInferenceConfig, BaseInferenceRunner
 from nemo.collections.tts.modules.magpietts_inference.utils import (
     ModelLoadConfig,
+    _add_common_args,
+    _add_easy_magpie_args,
+    _add_magpie_args,
+    _build_easy_magpie_config,
+    _build_magpie_config,
+    _configure_cuda_for_rank,
+    _get_torchrun_rank_info,
+    _merge_multiturn_rank_outputs,
+    _runner_eval_manifest_and_audio_dir,
+    _save_grouped_multiturn_filewise_metrics,
+    _select_runner_cls,
+    _wait_for_multiturn_rank_manifests,
+    append_metrics_to_csv,
+    create_formatted_metrics_mean_ci,
+    filter_datasets,
     get_experiment_name_from_checkpoint_path,
     load_easy_magpie_model,
     load_magpie_model,
     log_model_architecture_summary,
+    seed_all,
+    write_csv_header_if_needed,
 )
 from nemo.collections.tts.modules.magpietts_inference.visualization import create_combined_box_plot, create_violin_plot
-from nemo.collections.tts.modules.magpietts_modules import EOSDetectionMethod
 from nemo.utils import logging
-
-
-def parse_layer_list(layer_str: Optional[str]) -> Optional[List[int]]:
-    """Parse a comma-separated list of layer indices."""
-    if layer_str is None:
-        return None
-    return [int(l.strip()) for l in layer_str.split(",")]
-
-
-def write_csv_header_if_needed(csv_path: str, header: str) -> None:
-    """Write CSV header if file doesn't exist."""
-    if not os.path.exists(csv_path):
-        with open(csv_path, "w") as f:
-            f.write(header + "\n")
-
-
-def append_metrics_to_csv(csv_path: str, checkpoint_name: str, dataset: str, metrics: dict) -> None:
-    """Append metrics to a CSV file."""
-    values = [
-        checkpoint_name,
-        dataset,
-        metrics.get('cer_filewise_avg', ''),
-        metrics.get('wer_filewise_avg', ''),
-        metrics.get('cer_cumulative', ''),
-        metrics.get('wer_cumulative', ''),
-        metrics.get('ssim_pred_gt_avg', ''),
-        metrics.get('ssim_pred_context_avg', ''),
-        metrics.get('ssim_gt_context_avg', ''),
-        metrics.get('ssim_pred_gt_avg_alternate', ''),
-        metrics.get('ssim_pred_context_avg_alternate', ''),
-        metrics.get('ssim_gt_context_avg_alternate', ''),
-        metrics.get('esim_pred_gt_avg', ''),
-        metrics.get('ems_pred_gt_avg', ''),
-        metrics.get('cer_gt_audio_cumulative', ''),
-        metrics.get('wer_gt_audio_cumulative', ''),
-        metrics.get('utmosv2_avg', ''),
-        metrics.get('total_gen_audio_seconds', ''),
-        metrics.get('frechet_codec_distance', ''),
-        metrics.get('eou_cutoff_rate', ''),
-        metrics.get('eou_silence_rate', ''),
-        metrics.get('eou_noise_rate', ''),
-        metrics.get('eou_error_rate', ''),
-        metrics.get('katakana_cer_filewise_avg', ''),
-        metrics.get('katakana_cer_cumulative', ''),
-    ]
-    with open(csv_path, "a") as f:
-        f.write(",".join(str(v) for v in values) + "\n")
-    logging.info(f"Metrics appended to: {csv_path}")
-
-
-def _mean_finite(values: list):
-    vals = []
-    for value in values:
-        try:
-            value = float(value)
-        except (TypeError, ValueError):
-            continue
-        if np.isfinite(value):
-            vals.append(value)
-    return None if not vals else float(np.mean(vals))
-
-
-def _enrich_filewise_metrics_with_manifest(filewise_metrics: list, manifest_path: str) -> list:
-    """Attach multiturn manifest metadata back to evaluator filewise rows.
-
-    evaluate_generated_audio_dir() returns one filewise row per generated turn,
-    but the filtered row does not preserve source_sample_idx/turn_id. The
-    generated multiturn manifest has the same order as predicted_audio_*.wav, so
-    we merge by list index before grouping.
-    """
-    if manifest_path is None or not os.path.exists(manifest_path):
-        logging.warning(f"Could not enrich multiturn filewise metrics; manifest missing: {manifest_path}")
-        return filewise_metrics
-
-    manifest_records = read_manifest(manifest_path)
-    if len(manifest_records) != len(filewise_metrics):
-        logging.warning(
-            "Could not safely enrich multiturn filewise metrics; "
-            f"manifest rows={len(manifest_records)} filewise rows={len(filewise_metrics)} "
-            f"manifest_path={manifest_path}"
-        )
-        return filewise_metrics
-
-    enriched = []
-    for row, record in zip(filewise_metrics, manifest_records):
-        new_row = dict(row)
-        for key in [
-            "source_sample_idx",
-            "turn_id",
-            "speaker",
-            "rank",
-            "rank_local_idx",
-            "audio_filepath",
-            "context_audio_filepath",
-            "predicted_phoneme_text",
-            "predicted_phoneme_tokens",
-            "predicted_phoneme_token_labels",
-        ]:
-            if key in record and key not in new_row:
-                new_row[key] = record[key]
-        enriched.append(new_row)
-
-    return enriched
-
-
-def _group_multiturn_filewise_metrics_by_sample(filewise_metrics: list) -> list:
-    """Group turn-level multiturn metrics into one old-style row per sample.
-
-    Each grouped row keeps turn-by-turn CER/WER/SSIM/UTMOS/text/audio lists plus
-    averaged sample-level values. Rows are sorted by averaged CER descending so
-    the worst conversations/samples appear first.
-    """
-    grouped = {}
-
-    for row_idx, row in enumerate(filewise_metrics):
-        source_sample_idx = row.get("source_sample_idx", None)
-        if source_sample_idx is None:
-            source_sample_idx = row.get("speaker", None)
-        if source_sample_idx is None:
-            source_sample_idx = row_idx
-
-        key = str(source_sample_idx)
-        if key not in grouped:
-            grouped[key] = {
-                "source_sample_idx": source_sample_idx,
-                "speaker": row.get("speaker", source_sample_idx),
-                "rank": row.get("rank", None),
-                "target_audio_path": row.get("gt_audio_filepath", row.get("audio_filepath", "")),
-                "context_audio_path": row.get("context_audio_filepath", ""),
-                "turn_rows": [],
-            }
-        grouped[key]["turn_rows"].append(row)
-
-    grouped_rows = []
-    for _, group in grouped.items():
-        turns = group["turn_rows"]
-
-        def turn_sort_key(r):
-            try:
-                return int(r.get("turn_id", 0))
-            except (TypeError, ValueError):
-                return 0
-
-        turns = sorted(turns, key=turn_sort_key)
-
-        cer_turns = [r.get("cer") for r in turns]
-        wer_turns = [r.get("wer") for r in turns]
-        pred_context_ssim_turns = [r.get("pred_context_ssim") for r in turns]
-        pred_gt_ssim_turns = [r.get("pred_gt_ssim") for r in turns]
-        gt_context_ssim_turns = [r.get("gt_context_ssim") for r in turns]
-        pred_gt_esim_turns = [r.get("pred_gt_esim") for r in turns]
-        pred_gt_ems_turns = [r.get("pred_gt_ems") for r in turns]
-        utmosv2_turns = [r.get("utmosv2") for r in turns]
-        eou_type_turns = [r.get("eou_type") for r in turns]
-        eou_trailing_duration_turns = [r.get("eou_trailing_duration") for r in turns]
-        eou_trail_rms_ratio_turns = [r.get("eou_trail_rms_ratio") for r in turns]
-        predicted_phoneme_text_turns = [r.get("predicted_phoneme_text", "") for r in turns]
-        predicted_phoneme_tokens_turns = [r.get("predicted_phoneme_tokens", []) for r in turns]
-        predicted_phoneme_token_labels_turns = [r.get("predicted_phoneme_token_labels", []) for r in turns]
-
-        grouped_rows.append(
-            {
-                "source_sample_idx": group["source_sample_idx"],
-                "speaker": group["speaker"],
-                "rank": group["rank"],
-                "num_turns": len(turns),
-                # Sample-level averages over all turns.
-                "cer": _mean_finite(cer_turns),
-                "wer": _mean_finite(wer_turns),
-                "pred_context_ssim": _mean_finite(pred_context_ssim_turns),
-                "pred_gt_ssim": _mean_finite(pred_gt_ssim_turns),
-                "gt_context_ssim": _mean_finite(gt_context_ssim_turns),
-                "pred_gt_esim": _mean_finite(pred_gt_esim_turns),
-                "pred_gt_ems": _mean_finite(pred_gt_ems_turns),
-                "utmosv2": _mean_finite(utmosv2_turns),
-                "eou_trailing_duration": _mean_finite(eou_trailing_duration_turns),
-                "eou_trail_rms_ratio": _mean_finite(eou_trail_rms_ratio_turns),
-                # Turn-by-turn values, old-script style.
-                "turn_ids": [r.get("turn_id", i) for i, r in enumerate(turns)],
-                "cer_turns": cer_turns,
-                "wer_turns": wer_turns,
-                "pred_context_ssim_turns": pred_context_ssim_turns,
-                "pred_gt_ssim_turns": pred_gt_ssim_turns,
-                "gt_context_ssim_turns": gt_context_ssim_turns,
-                "pred_gt_esim_turns": pred_gt_esim_turns,
-                "pred_gt_ems_turns": pred_gt_ems_turns,
-                "utmosv2_turns": utmosv2_turns,
-                "eou_type_turns": eou_type_turns,
-                "eou_trailing_duration_turns": eou_trailing_duration_turns,
-                "eou_trail_rms_ratio_turns": eou_trail_rms_ratio_turns,
-                "predicted_phoneme_text_turns": predicted_phoneme_text_turns,
-                "predicted_phoneme_tokens_turns": predicted_phoneme_tokens_turns,
-                "predicted_phoneme_token_labels_turns": predicted_phoneme_token_labels_turns,
-                "reference_text": [r.get("gt_text", "") for r in turns],
-                "asr_hyp": [r.get("pred_text", "") for r in turns],
-                "pred_audio_paths": [r.get("pred_audio_filepath", "") for r in turns],
-                "target_audio_path": group["target_audio_path"],
-                "context_audio_path": group["context_audio_path"],
-                "turn_metrics": turns,
-            }
-        )
-
-    grouped_rows.sort(
-        key=lambda r: (
-            r.get("cer") is not None,
-            float(r["cer"]) if r.get("cer") is not None else -1.0,
-        ),
-        reverse=True,
-    )
-    return grouped_rows
-
-
-def _write_grouped_multiturn_filewise_metrics_csv(csv_path: str, grouped_rows: list) -> None:
-    fieldnames = [
-        "source_sample_idx",
-        "speaker",
-        "rank",
-        "num_turns",
-        "cer",
-        "wer",
-        "pred_context_ssim",
-        "pred_gt_ssim",
-        "gt_context_ssim",
-        "pred_gt_esim",
-        "pred_gt_ems",
-        "utmosv2",
-        "eou_trailing_duration",
-        "eou_trail_rms_ratio",
-        "turn_ids",
-        "cer_turns",
-        "wer_turns",
-        "pred_context_ssim_turns",
-        "pred_gt_ssim_turns",
-        "gt_context_ssim_turns",
-        "pred_gt_esim_turns",
-        "pred_gt_ems_turns",
-        "utmosv2_turns",
-        "eou_type_turns",
-        "eou_trailing_duration_turns",
-        "eou_trail_rms_ratio_turns",
-        "target_audio_path",
-        "context_audio_path",
-        "pred_audio_paths",
-        "reference_text",
-        "asr_hyp",
-        "predicted_phoneme_text_turns",
-        "predicted_phoneme_tokens_turns",
-        "predicted_phoneme_token_labels_turns",
-    ]
-
-    def csv_value(value):
-        if isinstance(value, (list, dict)):
-            value = json.dumps(value, ensure_ascii=False)
-        if value is None:
-            value = ""
-        value = str(value).replace('"', '""')
-        if "," in value or "\n" in value or "[" in value or "{" in value:
-            value = f'"{value}"'
-        return value
-
-    with open(csv_path, "w", encoding="utf-8") as f:
-        f.write(",".join(fieldnames) + "\n")
-        for row in grouped_rows:
-            f.write(",".join(csv_value(row.get(k, "")) for k in fieldnames) + "\n")
-
-
-def _save_grouped_multiturn_filewise_metrics(
-    eval_dir: str,
-    dataset: str,
-    repeat_idx: int,
-    filewise_metrics: list,
-    manifest_path: str,
-) -> None:
-    enriched_filewise = _enrich_filewise_metrics_with_manifest(filewise_metrics, manifest_path)
-    grouped_rows = _group_multiturn_filewise_metrics_by_sample(enriched_filewise)
-
-    json_path = os.path.join(eval_dir, f"{dataset}_grouped_filewise_metrics_{repeat_idx}.json")
-    csv_path = os.path.join(eval_dir, f"{dataset}_grouped_filewise_metrics_{repeat_idx}.csv")
-
-    with open(json_path, "w", encoding="utf-8") as f:
-        json.dump(grouped_rows, f, indent=4, ensure_ascii=False)
-
-    _write_grouped_multiturn_filewise_metrics_csv(csv_path, grouped_rows)
-
-    logging.info(f"Saved grouped multiturn filewise metrics JSON to: {json_path}")
-    logging.info(f"Saved grouped multiturn filewise metrics CSV to: {csv_path}")
-
-
-def create_formatted_metrics_mean_ci(metrics_mean_ci: dict) -> dict:
-    """Create formatted metrics mean CI."""
-    for k, v in metrics_mean_ci.items():
-        if isinstance(v, list):
-            mean, ci = float(v[0]), float(v[1])
-            logging.info(f"Metric {k}: {mean:.4f} ± {ci:.4f}")
-            metrics_mean_ci[k] = f"{mean:.4f} ± {ci:.4f}"
-    return metrics_mean_ci
-
-
-def filter_datasets(
-    dataset_meta_info: dict,
-    datasets: Optional[List[str]],
-) -> List[str]:
-    """Select datasets from the dataset meta info."""
-    if datasets is None:
-        # Dataset filtering not specified, return all datasets.
-        return list(dataset_meta_info.keys())
-    else:
-        datasets = datasets.split(",")
-        # Check if requested datasets are valid.
-        for dataset in datasets:
-            if dataset not in dataset_meta_info:
-                raise ValueError(f"Dataset {dataset} not found in dataset meta info")
-        # Return all requested datasets.
-        return datasets
-
-
-def _runner_eval_manifest_and_audio_dir(runner: BaseInferenceRunner, default_manifest: str, default_audio_dir: str):
-    """Return evaluation manifest/audio dir produced by the runner, if any."""
-    eval_manifest = getattr(runner, "evaluation_manifest_path", None) or default_manifest
-    eval_audio_dir = getattr(runner, "evaluation_audio_dir", None) or default_audio_dir
-    return eval_manifest, eval_audio_dir
-
-
-def _get_torchrun_rank_info() -> Tuple[int, int, int]:
-    """Return (rank, world_size, local_rank) from torchrun/SLURM env vars.
-
-    We intentionally do not initialize torch.distributed here. The inference
-    script only needs env-based sharding, while NeMo evaluation models can run
-    without distributed collectives.
-    """
-    rank = int(os.environ.get("RANK", os.environ.get("SLURM_PROCID", "0")))
-    world_size = int(os.environ.get("WORLD_SIZE", "1"))
-    local_rank = int(os.environ.get("LOCAL_RANK", os.environ.get("SLURM_LOCALID", "0")))
-    return rank, world_size, local_rank
-
-
-def _configure_cuda_for_rank() -> Tuple[int, int, int]:
-    rank, world_size, local_rank = _get_torchrun_rank_info()
-    if torch.cuda.is_available():
-        device_count = torch.cuda.device_count()
-        if device_count > 0:
-            torch.cuda.set_device(local_rank % device_count)
-            logging.info(
-                f"Using CUDA device {local_rank % device_count}; "
-                f"rank={rank}, local_rank={local_rank}, world_size={world_size}"
-            )
-    return rank, world_size, local_rank
-
-
-def _wait_for_multiturn_rank_manifests(repeat_audio_dir: str, world_size: int, timeout_sec: int = 7200) -> None:
-    deadline = time.time() + timeout_sec
-    while time.time() < deadline:
-        missing = []
-        for rank in range(world_size):
-            path = os.path.join(
-                repeat_audio_dir,
-                f"rank_{rank:04d}",
-                f"multiturn_user_audio_turn_manifest_rank{rank:04d}.jsonl",
-            )
-            if not os.path.exists(path):
-                missing.append(path)
-        if not missing:
-            return
-        time.sleep(5)
-    raise RuntimeError(f"Timed out waiting for multiturn rank manifests: {missing}")
-
-
-def _copy_or_link(src: str, dst: str, required: bool = False) -> None:
-    if src is None or not os.path.exists(src):
-        if os.path.lexists(dst):
-            os.remove(dst)
-        if required:
-            raise FileNotFoundError(f"Missing required merge source: {src} -> {dst}")
-        return
-
-    os.makedirs(os.path.dirname(dst), exist_ok=True)
-
-    if os.path.lexists(dst):
-        os.remove(dst)
-
-    # Prefer real files for evaluator inputs; broken symlinks confuse librosa/UTMOS.
-    shutil.copyfile(src, dst)
-
-
-def _merge_multiturn_rank_outputs(repeat_audio_dir: str, world_size: int, save_predicted_codes: bool) -> str:
-    """Merge rank-local multiturn outputs into one EasyMagpie-compatible dir.
-
-    Each rank writes local files named predicted_audio_0.wav, target_audio_0.wav,
-    context_audio_0.wav, predicted_codes_0.pt, ... inside rank_XXXX/. This
-    function remaps them to contiguous global indices in repeat_audio_dir/ and
-    writes a merged turn-level manifest.
-    """
-    # clean previous merged files
-    for pattern in [
-        "predicted_audio_*.wav",
-        "target_audio_*.wav",
-        "context_audio_*.wav",
-        "predicted_codes_*.pt",
-    ]:
-        for path in Path(repeat_audio_dir).glob(pattern):
-            if path.is_symlink() or path.exists():
-                path.unlink(missing_ok=True)
-
-    merged_records = []
-    global_idx = 0
-
-    for rank in range(world_size):
-        rank_dir = os.path.join(repeat_audio_dir, f"rank_{rank:04d}")
-        rank_manifest = os.path.join(rank_dir, f"multiturn_user_audio_turn_manifest_rank{rank:04d}.jsonl")
-        if not os.path.exists(rank_manifest):
-            raise FileNotFoundError(f"Missing rank manifest: {rank_manifest}")
-
-        with open(rank_manifest, "r", encoding="utf-8") as f:
-            rank_records = [json.loads(line) for line in f if line.strip()]
-
-        for local_idx, record in enumerate(rank_records):
-            pred_src = os.path.join(rank_dir, f"predicted_audio_{local_idx}.wav")
-            pred_dst = os.path.join(repeat_audio_dir, f"predicted_audio_{global_idx}.wav")
-            _copy_or_link(pred_src, pred_dst, required=True)
-
-            if save_predicted_codes:
-                code_src = os.path.join(rank_dir, f"predicted_codes_{local_idx}.pt")
-                code_dst = os.path.join(repeat_audio_dir, f"predicted_codes_{global_idx}.pt")
-                _copy_or_link(code_src, code_dst, required=False)
-
-            target_src = os.path.join(rank_dir, record.get("audio_filepath", f"target_audio_{local_idx}.wav"))
-            target_dst = os.path.join(repeat_audio_dir, f"target_audio_{global_idx}.wav")
-            _copy_or_link(target_src, target_dst, required=True)
-
-            context_src = os.path.join(
-                rank_dir,
-                record.get("context_audio_filepath", f"context_audio_{local_idx}.wav"),
-            )
-            context_dst = os.path.join(repeat_audio_dir, f"context_audio_{global_idx}.wav")
-            _copy_or_link(context_src, context_dst, required=True)
-
-            merged = dict(record)
-            merged["audio_filepath"] = f"target_audio_{global_idx}.wav"
-            merged["context_audio_filepath"] = f"context_audio_{global_idx}.wav"
-            merged["rank"] = rank
-            merged["rank_local_idx"] = local_idx
-            merged_records.append(merged)
-            global_idx += 1
-
-    merged_manifest = os.path.join(repeat_audio_dir, "multiturn_user_audio_turn_manifest.jsonl")
-    with open(merged_manifest, "w", encoding="utf-8") as f:
-        for record in merged_records:
-            f.write(json.dumps(record, ensure_ascii=False) + "\n")
-
-    logging.info(f"Merged {len(merged_records)} multiturn turn records into {merged_manifest}")
-    return merged_manifest
 
 
 def run_inference_and_evaluation(
@@ -849,277 +399,6 @@ def run_inference_and_evaluation(
     return None, None
 
 
-def _get_shared_inference_param_names() -> set:
-    """Return the field names shared by ModelInferenceParameters and EasyModelInferenceParameters."""
-    magpie_fields = {f.name for f in fields(ModelInferenceParameters)}
-    easy_fields = {f.name for f in fields(EasyModelInferenceParameters)}
-    return magpie_fields & easy_fields
-
-
-def _add_inference_param_fields(
-    group: argparse._ArgumentGroup,
-    param_cls: type,
-    skip_fields: Optional[set] = None,
-    only_fields: Optional[set] = None,
-) -> None:
-    """Auto-generate argparse arguments from fields of a dataclass.
-
-    Args:
-        group: The argparse argument group to add arguments to.
-        param_cls: The dataclass whose fields to add.
-        skip_fields: Field names to skip (already added by another group).
-        only_fields: If provided, only add fields whose names are in this set.
-    """
-    if skip_fields is None:
-        skip_fields = set()
-    for f in fields(param_cls):
-        if f.name in skip_fields:
-            continue
-        if only_fields is not None and f.name not in only_fields:
-            continue
-        extra_args: dict = {"type": f.type}
-        if f.type == bool:
-            extra_args = {"action": "store_true"}
-        if f.name in ("estimate_alignment_from_layers", "apply_prior_to_layers"):
-            extra_args = {
-                "help": "Must be a comma separate string. Not enclosed in brackets",
-                "type": str,
-            }
-        elif f.name == "eos_detection_method":
-            extra_args["choices"] = [m.value for m in EOSDetectionMethod]
-        group.add_argument(f"--{f.name}", **extra_args)
-
-
-def _add_common_args(parser: argparse.ArgumentParser) -> None:
-    """Add arguments shared by all model types."""
-
-    parser.add_argument(
-        '--model_type',
-        type=str,
-        default='magpie',
-        choices=['magpie', 'easy_magpie'],
-        help='Model type: "magpie" for encoder-decoder MagpieTTSModel, '
-        '"easy_magpie" for decoder-only EasyMagpieTTSInferenceModel',
-    )
-    parser.add_argument(
-        '--deterministic',
-        action='store_true',
-        help='Attempts to make results deterministic to the best that can be done. Used for testing',
-    )
-
-    # Model loading
-    model_group = parser.add_argument_group('Model Loading')
-    model_group.add_argument(
-        '--hparams_files',
-        type=str,
-        default=None,
-        help='Comma-separated paths to hparams.yaml files (use with --checkpoint_files)',
-    )
-    model_group.add_argument(
-        '--checkpoint_files',
-        type=str,
-        default=None,
-        help='Comma-separated paths to .ckpt files (use with --hparams_files)',
-    )
-    model_group.add_argument(
-        '--nemo_files',
-        type=str,
-        default=None,
-        help='Comma-separated paths to .nemo files (alternative to hparams + checkpoint)',
-    )
-    model_group.add_argument(
-        '--codecmodel_path',
-        type=str,
-        required=True,
-        help='Path to the audio codec model',
-    )
-    model_group.add_argument(
-        '--hparams_file_from_wandb',
-        action='store_true',
-        help='Set if hparams file was exported from wandb',
-    )
-    model_group.add_argument(
-        '--legacy_codebooks',
-        action='store_true',
-        help='Use legacy codebook indices (for old checkpoints)',
-    )
-    model_group.add_argument(
-        '--legacy_text_conditioning',
-        action='store_true',
-        help='Use legacy text conditioning (for old checkpoints)',
-    )
-
-    # Dataset and output
-    data_group = parser.add_argument_group('Dataset and Output')
-    data_group.add_argument(
-        '--datasets_json_path',
-        type=str,
-        required=True,
-        default=None,
-        help='Path to dataset configuration JSON file',
-    )
-    data_group.add_argument(
-        '--datasets_base_path',
-        type=Path,
-        default=None,
-        help='Optional base path that paths in the "datasets_json_path" file are relative to',
-    )
-    data_group.add_argument(
-        '--datasets',
-        type=str,
-        default=None,
-        help='Comma-separated list of dataset names to process',
-    )
-    data_group.add_argument(
-        '--tokenizer_name',
-        type=str,
-        default="english_phoneme",
-        help='Default tokenizer to use when a language or dataset specific tokenizer is not provided.',
-    )
-    data_group.add_argument('--out_dir', type=str, required=True, help='Output directory')
-    data_group.add_argument('--log_exp_name', action='store_true')
-    data_group.add_argument('--clean_up_disk', action='store_true')
-
-    # Common inference parameters
-    infer_group = parser.add_argument_group('Common Inference Parameters')
-    infer_group.add_argument('--batch_size', type=int, default=32)
-    infer_group.add_argument('--use_cfg', action='store_true', help='Enable classifier-free guidance')
-    infer_group.add_argument('--use_local_transformer', action='store_true')
-
-    # Model inference parameters shared by both MagpieTTS and EasyMagpieTTS
-    shared_param_names = _get_shared_inference_param_names()
-    _add_inference_param_fields(infer_group, ModelInferenceParameters, only_fields=shared_param_names)
-
-    # Evaluation
-    eval_group = parser.add_argument_group('Evaluation')
-    eval_group.add_argument('--run_evaluation', action='store_true', help='Run evaluation after inference')
-    eval_group.add_argument('--sv_model', type=str, default="titanet", choices=["titanet", "wavlm"])
-    eval_group.add_argument(
-        '--asr_model_name',
-        type=str,
-        default='nvidia/parakeet-tdt-1.1b',
-        help="ASR model to use for WER calculation, when not provided in dataset config",
-    )
-    eval_group.add_argument(
-        '--asr_model_type',
-        type=str,
-        default='nemo',
-        choices=['nemo', 'nemo_with_prompt', 'whisper'],
-        help="Type of ASR model provided in 'asr_model_name'",
-    )
-    eval_group.add_argument(
-        '--language', type=str, default="en", help='Language to use, when not provided in dataset config'
-    )
-    eval_group.add_argument(
-        '--eou_model_name',
-        type=str,
-        default="facebook/wav2vec2-base-960h",
-        help=(
-            'Hugging Face model id or local path to the EoU wav2vec2 model directory. '
-            'For offline use, download the model locally and pass the directory path here.'
-        ),
-    )
-    eval_group.add_argument('--num_repeats', type=int, default=1)
-    eval_group.add_argument('--confidence_level', type=float, default=0.95)
-    eval_group.add_argument('--disable_utmosv2', action='store_true')
-    eval_group.add_argument('--with_emotion_metrics', action='store_true')
-    eval_group.add_argument(
-        '--strip_text_annotations_for_metrics',
-        action='store_true',
-        help='Strip bracket/tag/control annotations from reference and ASR hypothesis text while computing text metrics.',
-    )
-    eval_group.add_argument('--emotion_model_size', type=str, default="small", choices=["small", "large"])
-    eval_group.add_argument(
-        '--emotion_embedding_type',
-        type=str,
-        default="score_vector",
-        choices=["head_concat", "head_mean", "score_vector"],
-    )
-    eval_group.add_argument('--emotion_cache_dir', type=str, default=None)
-    eval_group.add_argument(
-        '--violin_plot_metrics',
-        type=str,
-        nargs='*',
-        default=['cer', 'pred_context_ssim', 'utmosv2'],
-    )
-    eval_group.add_argument('--disable_fcd', action='store_true')
-    eval_group.add_argument("--asr_batch_size", type=int, default=32)
-    eval_group.add_argument("--eou_batch_size", type=int, default=32)
-
-    # Quality targets
-    target_group = parser.add_argument_group('Quality Targets')
-    target_group.add_argument('--cer_target', type=float, default=None)
-    target_group.add_argument('--ssim_target', type=float, default=None)
-
-
-def seed_all(seed: int):
-    """
-    Attempts to make script deterministic
-    """
-    torch.manual_seed(seed)
-    random.seed(seed)
-    np.random.seed(seed)
-    torch.backends.cudnn.benchmark = False
-    torch.use_deterministic_algorithms(True)
-
-
-def _add_magpie_args(parser: argparse.ArgumentParser) -> None:
-    """Add arguments specific to encoder-decoder MagpieTTSModel."""
-    group = parser.add_argument_group('MagpieTTS-specific Parameters')
-
-    # MagpieTTS-specific model inference parameters (attention prior, EOS, etc.)
-    shared_param_names = _get_shared_inference_param_names()
-    _add_inference_param_fields(group, ModelInferenceParameters, skip_fields=shared_param_names)
-
-    group.add_argument('--maskgit_n_steps', type=int, default=3)
-    group.add_argument('--maskgit_noise_scale', type=float, default=0.0)
-    group.add_argument('--maskgit_fixed_schedule', type=int, nargs='+', default=None)
-    group.add_argument(
-        '--maskgit_sampling_type',
-        default=None,
-        choices=["default", "causal", "purity_causal", "purity_default"],
-    )
-
-
-def _add_easy_magpie_args(parser: argparse.ArgumentParser) -> None:
-    """Add arguments specific to decoder-only EasyMagpieTTSInferenceModel."""
-    group = parser.add_argument_group('EasyMagpieTTS-specific Parameters')
-    group.add_argument(
-        '--easy_magpie_inference_mode',
-        type=str,
-        default='single_turn',
-        choices=['single_turn', 'multiturn_user_audio'],
-    )
-    group.add_argument('--max_eval_turns', type=int, default=6)
-    group.add_argument('--no_save_debug_multiturn_audio', action='store_true')
-    group.add_argument(
-        '--phoneme_input_type',
-        type=str,
-        default='gt',
-        choices=['gt', 'predicted'],
-        help='Source of phoneme input for decoder-only model',
-    )
-    group.add_argument(
-        '--phoneme_sampling_method',
-        type=str,
-        default='argmax',
-        choices=['argmax', 'multinomial'],
-        help='Sampling method for phoneme prediction',
-    )
-    group.add_argument('--dropout_text_input', action='store_true', help='Force dropout on text input')
-    group.add_argument(
-        '--phoneme_tokenizer_path',
-        type=str,
-        default=None,
-        help='Override path to the phoneme tokenizer file (overrides the path stored in the checkpoint config)',
-    )
-    group.add_argument(
-        '--disable_cas_for_context_text',
-        action='store_true',
-        help='Skip CAS embeddings for context text when loading legacy EasyMagpieTTS models',
-    )
-
-
 def create_argument_parser() -> argparse.ArgumentParser:
     """Create the CLI argument parser with all argument groups."""
     parser = argparse.ArgumentParser(
@@ -1131,68 +410,6 @@ def create_argument_parser() -> argparse.ArgumentParser:
     _add_magpie_args(parser)
     _add_easy_magpie_args(parser)
     return parser
-
-
-def _build_inference_params_from_args(param_cls: type, args):
-    """Extract inference parameters from parsed CLI args for the given dataclass."""
-    params = {}
-    for f in fields(param_cls):
-        arg_val = vars(args).get(f.name)
-        if arg_val is not None:
-            if f.name in ("estimate_alignment_from_layers", "apply_prior_to_layers"):
-                params[f.name] = parse_layer_list(arg_val)
-            else:
-                params[f.name] = arg_val
-    return param_cls.from_dict(params)
-
-
-def _build_magpie_config(args) -> MagpieInferenceConfig:
-    return MagpieInferenceConfig(
-        model_inference_parameters=_build_inference_params_from_args(ModelInferenceParameters, args),
-        batch_size=args.batch_size,
-        use_cfg=args.use_cfg,
-        apply_attention_prior=args.apply_attention_prior,
-        use_local_transformer=args.use_local_transformer,
-        maskgit_n_steps=args.maskgit_n_steps,
-        maskgit_noise_scale=args.maskgit_noise_scale,
-        maskgit_fixed_schedule=args.maskgit_fixed_schedule,
-        maskgit_sampling_type=args.maskgit_sampling_type,
-        default_tokenizer_name=args.tokenizer_name,
-    )
-
-
-def _build_easy_magpie_config(args) -> EasyMagpieInferenceConfig:
-    cfg_cls = (
-        EasyMagpieMultiturnUserAudioInferenceConfig
-        if args.easy_magpie_inference_mode == 'multiturn_user_audio'
-        else EasyMagpieInferenceConfig
-    )
-    kwargs = dict(
-        model_inference_parameters=_build_inference_params_from_args(EasyModelInferenceParameters, args),
-        batch_size=args.batch_size,
-        use_cfg=args.use_cfg,
-        use_local_transformer=args.use_local_transformer,
-        phoneme_input_type=args.phoneme_input_type,
-        phoneme_sampling_method=args.phoneme_sampling_method,
-        dropout_text_input=args.dropout_text_input,
-        default_tokenizer_name=args.tokenizer_name,
-    )
-    if cfg_cls is EasyMagpieMultiturnUserAudioInferenceConfig:
-        kwargs.update(
-            max_eval_turns=args.max_eval_turns,
-            save_debug_multiturn_audio=not args.no_save_debug_multiturn_audio,
-        )
-    return cfg_cls(**kwargs)
-
-
-def _select_runner_cls(args):
-    if args.model_type == 'magpie':
-        if args.easy_magpie_inference_mode != 'single_turn':
-            raise ValueError('--easy_magpie_inference_mode is only supported with --model_type easy_magpie')
-        return MagpieInferenceRunner
-    if args.easy_magpie_inference_mode == 'multiturn_user_audio':
-        return EasyMagpieMultiturnUserAudioInferenceRunner
-    return EasyMagpieInferenceRunner
 
 
 def main(argv=None):
